fixed testing for getaction

+++ models/action.go
@@ -2,21 +2,25 @@ package models
 
 //ActionInput represents information needed for each action input
 type ActionInput struct {
-	Action string `json:"action"` // the name of the action
-	Time   uint   `json:"time"`   // time taken for the action
+	Action string  `json:"action"` // the name of the action
+	Time   float64 `json:"time"`   // time taken for the action
 }
 
 //ActionOutput represents information needed for each action output
 type ActionOutput struct {
-	Action string `json:"action"` // name of the action
-	Avg    uint   `json:"avg"`    // average time taken for each action
+	Action string  `json:"action"` // name of the action
+	Avg    float64 `json:"avg"`    // average time taken for each action
 }
 
+//ActionCounter represents information needed to keep track how many times
+// an action is called
 type ActionCounter struct {
-	TotalTime uint
-	Counter   uint
+	TotalTime float64 // total time for each action
+	Counter   uint    // counter for the number of times an action is called.
 }
 
+//ActionMap stores the action along with its time and counter
+// key - action: value - ActionCounter struct
 type ActionMap map[string]ActionCounter
 
 // Validate validates the field for ActionInput struct
